rename: 서비스 로직 파일 구조 변경
